Fix dividend calculation to not include 'voided' transactions
In the get_total_dividends() method, when we check each item in data, we simply use item['account'] in the sum. However, that can result in over-counting in the case where you have transactions with 'voided' state. There could be other such states, and hence the safest way is to only include states that are known to be yielding the right dividend transactions, i.e 'paid' and 'reinvested'.