Allow the print() output to go to anywhere
Rather than always outputting to stdout, I added a new set_output
and get_output in helper.py that allows the client to set the stream
that all print statements will go to.  It defaults to stdout so the
standard behavior should remain unchanged.

Rather than pass this around everywhere, I made the stream a global.
I stayed with the default print() function and just passed in the
stream with the file= keyword.  It may have made more sense to create
a new print function wrapper. So that the file= logic is in one place

The other controversial part is the write_spinner function.  I did the
simple approach to do the spinner only if the output is equal to stdout.
Otherwise the spinner is ignored.  I think that's fine for most use
cases.  I could see a case where a unix user is redirecting the
output of to a file, In that case, the stdout check will pass and all
of the \b will be put into the file.  It probably makes more sense to
do the spinner only if we're going to a terminal, but I'm not going
that far.